C++: rename headers from .h to .hpp
