+++ swig/cpp/examples/cpp_application_changes_example.cpp
@@ -26,7 +26,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-#include "Session.h"
+#include "Session.hpp"
 
 #define MAX_LEN 100
