+++ swig/cpp/examples/cpp_application_example.cpp
@@ -25,7 +25,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-#include "Session.h"
+#include "Session.hpp"
 
 #define MAX_LEN 100
