+++ swig/cpp/examples/cpp_get_item_example.cpp
@@ -22,7 +22,7 @@
 #include <iostream>
 #include <memory>
 
-#include "Session.h"
+#include "Session.hpp"
 
 using namespace std;
