+++ swig/cpp/examples/cpp_get_items_iter_example.cpp
@@ -21,7 +21,7 @@
 
 #include <iostream>
 
-#include "Session.h"
+#include "Session.hpp"
 
 using namespace std;
