+++ swig/cpp/examples/cpp_rpc_example.cpp
@@ -27,7 +27,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-#include "Session.h"
+#include "Session.hpp"
 
 #define MAX_LEN 100
