+++ swig/cpp/src/Internal.hpp
