+++ swig/cpp/src/Struct.hpp
@@ -25,8 +25,8 @@
 #include <iostream>
 #include <memory>
 
-#include "Sysrepo.h"
-#include "Internal.h"
+#include "Sysrepo.hpp"
+#include "Internal.hpp"
 
 extern "C" {
 #include "sysrepo.h"
