+++ swig/cpp/src/Xpath.hpp
