+++ swig/cpp/tests/changes.cpp
@@ -4,7 +4,7 @@
 #include <cstring>
 #include <unistd.h>
 
-#include "Session.h"
+#include "Session.hpp"
 
 #define MAX_LEN 100
