+++ swig/cpp/tests/operations.cpp
@@ -4,7 +4,7 @@
 #include <cstring>
 #include <unistd.h>
 
-#include "Session.h"
+#include "Session.hpp"
 
 using namespace std;
