+++ swig/lua/libsysrepoLua.i
@@ -27,9 +27,9 @@
 #include <vector>
 #include <memory>
 
-#include "Sysrepo.h"
-#include "Struct.h"
-#include "Session.h"
+#include "Sysrepo.hpp"
+#include "Struct.hpp"
+#include "Session.hpp"
 
 class Callback_lua {
 public:
